DynamoDB Limit Support
We add a limit parameter to the ScanRequest or QueryRequest, if the constraints  object has a limit.
When the DDB client execs a Scan or Query, it will cap the results to the limit value.
Then, when we parse the iterator, it will have no more than that limit number of items.

Case 1: Limit is < Batch size of a ddb scan/query
  - ddb scan/query will return exactly limit values. Iterator will say true to hasNext() limit times.
  our logic tells the iterator consumer to stop reading once we hit limit values. we stop.
Case 2: Limit is > Batch size of a ddb scan / query
  - ddb scan/query returns max values  < limit. Iterator says true to hasNext() max times. our logic sees
  max < limit, so it asks the iterator for hasNext(). Iterator sees it ran out  of values, and conducts
  another batch scan/query. We repeat until we hit the limit or run out of rows.